feat: portfolio polish — logrolling proof, CI, types, README front page
Turn the working v0 into a portfolio-grade standalone repo. Substance and
presentation, guided by a 4-lens review (senior-engineer, negotiation-science,
python-maintainer, DevRel).

Substance
- baseline.py: uniform price-split control + head-to-head comparison. Proves the
  headline claim — at matched buyer utility, logrolling leaves the supplier +0.27
  utility better off than splitting the difference. Runnable via `neg-sim
  --baseline`; asserted in test_baseline.py (fails CI if the edge collapses).
- Replaces the old near-tie logrolling assertion with the clean, un-confounded
  baseline proof.
- Golden reference-metrics fixture + regression test so README numbers can't drift.

Engineering polish
- mypy --strict passes; removed the loop.py `type: ignore` by threading a shared
  Status literal; added py.typed.
- GitHub Actions CI: ruff check + format, mypy, pytest w/ 90% coverage gate,
  across ubuntu+windows × py3.11/3.12/3.13. Coverage is 95.7%.
- MIT LICENSE (was "Proprietary"); classifiers, keywords, project URLs, pydantic
  upper bound for reproducible eval numbers.

Presentation
- neg-sim gains --baseline and --transcript (turn-by-turn replay); main() takes
  argv for testability.
- README front page: zero-authority hook, real demo output, a live transcript,
  the baseline proof table, two Mermaid diagrams (decide() FSM + fill_package
  pipeline), an engineering-judgment section (Boulware bug, numeric guard,
  determinism), honest eval framing, and negotiation-theory citations.
- Persona docstrings mapped to procurement archetypes; BATNA/ZOPA vocabulary.
- Untracked the .claude/ toolkit (kept on disk) so the 13 source files are the
  repo's center of gravity.

58 tests green · mypy strict clean · ruff clean · 95.7% coverage

+++ src/negotiation_agent/engine.py
@@ -104,7 +104,7 @@ class DealEngine:
         t = min(max(round_index, 0), cfg.max_rounds)
         frac = t / cfg.max_rounds
         span = env.target_utility - env.reservation_utility
-        return env.reservation_utility + span * (1.0 - frac**cfg.beta)
+        return float(env.reservation_utility + span * (1.0 - frac**cfg.beta))
 
     # ---- Decision ----------------------------------------------------------
 
@@ -228,14 +228,14 @@ class DealEngine:
         except InfeasiblePackage:
             # Caps have ratcheted below what theta needs — nothing legal left to offer.
             return (
-                self._escalate(state.round_index, theta, reason="no_feasible_counter",
-                               incoming_utility=u_in),
+                self._escalate(
+                    state.round_index, theta, reason="no_feasible_counter", incoming_utility=u_in
+                ),
                 state,
             )
 
         new_caps = {
-            name: min(state.concession_caps.get(name, 1.0), planned_v[name])
-            for name in planned_v
+            name: min(state.concession_caps.get(name, 1.0), planned_v[name]) for name in planned_v
         }
         decision = EngineDecision(
             outcome=Outcome.COUNTER,
